% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow functions.R
\name{regional_SWE}
\alias{regional_SWE}
\title{Regional SWE plot}
\usage{
regional_SWE(
  x,
  y,
  currentyear,
  act,
  surface,
  plot_width = 20,
  plot_height = 12,
  plot_name = "Regional SWE"
)
}
\arguments{
\item{x}{The dataframe containing snow survey data.}

\item{y}{The site metadata sheet.}

\item{currentyear}{The current year, will be highlighted in red}

\item{act}{The activity of the snow survey site. Options are "A" for active and "IA" for inactive.}

\item{surface}{The surface type for the snow survey. Options are "upland" or "lake."}

\item{plot_width}{Default set to 20}

\item{plot_height}{Default set to 12}

\item{plot_name}{Saves a png called "Regional SWE" in the working directory}
}
\description{
Creates boxplot for SWE based on location; South Slave, North Slave, Inuvialuit/Gwich'in, Sahtu and Dehcho regions.
}
