% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Snow functions.R
\name{SummaryTableSO}
\alias{SummaryTableSO}
\title{Spring outlook summary table}
\usage{
SummaryTableSO(x, surface, minyear, maxyear, curmaxyear, act, filepath)
}
\arguments{
\item{x}{Snow survey dataframe}

\item{surface}{Filters for surface type. Specify "upland", "lake" or NULL (all surface types).}

\item{minyear}{Minimum calculation year (minus 1) for the normal period. For example, for a 2001-2020 normal period minyear is set to 2000.}

\item{maxyear}{Maximum calculation year (plus 1) for the normal period. For example, for a 2001-2020 normal period maxyear is set to 2021.}

\item{curmaxyear}{Current calendar year.}

\item{act}{Set to either "A" or "IA." Filters for active snow survey sites.}

\item{filepath}{Specify filepath for output excel file including desired file name. For example "~/SummaryTableSO.xlsx"}
}
\description{
Takes snow survey values from the SnowDB database and generates a summary table for the current year.
Output is an excel file written as "filepath" argument.
swe, depth and site variables must be named "SWE", "depth" and "site" in the input dataframe.
}
